+++ bigmler/analyze_options.py
@@ -39,7 +39,7 @@ def subparser_options(subparsers, defaults={}, constants={}):
 
     parser_analysis = subparsers.add_parser('analyze')
 
-    common_options(parser_analysis, defaults=defaults)
+    common_options(parser_analysis, defaults=defaults, constants=constants);
 
     # K-fold cross-validation flag
     parser_analysis.add_argument('--cross-validation',