+++ bigmler/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-__version__ = '1.8.12'
+__version__ = '1.9.0'
 
 if __name__ == '__main__':
     import sys